Refactor EmailService request
