+++ src/Http/Controllers/EmailServices/EmailServicesController.php
@@ -9,7 +9,7 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\View\View;
 use Sendportal\Base\Http\Controllers\Controller;
-use Sendportal\Base\Http\Requests\EmailServiceStoreRequest;
+use Sendportal\Base\Http\Requests\EmailServiceRequest;
 use Sendportal\Base\Http\Requests\EmailServiceUpdateRequest;
 use Sendportal\Base\Repositories\EmailServiceTenantRepository;
 
@@ -43,7 +43,7 @@ class EmailServicesController extends Controller
     /**
      * @throws Exception
      */
-    public function store(EmailServiceStoreRequest $request): RedirectResponse
+    public function store(EmailServiceRequest $request): RedirectResponse
     {
         $emailServiceType = $this->emailServices->findType($request->type_id);
 
@@ -73,7 +73,7 @@ class EmailServicesController extends Controller
     /**
      * @throws Exception
      */
-    public function update(EmailServiceUpdateRequest $request, int $emailServiceId): RedirectResponse
+    public function update(EmailServiceRequest $request, int $emailServiceId): RedirectResponse
     {
         $emailService = $this->emailServices->find(auth()->user()->currentWorkspace()->id, $emailServiceId, ['type']);
