+++ src/Models/EmailServiceType.php
@@ -28,80 +28,4 @@ class EmailServiceType extends BaseModel
     {
         return static::$types[$typeId] ?? null;
     }
-
-    public static function resolveValidationRules($typeId): array
-    {
-        switch ($typeId) {
-            case self::SES:
-                return [
-                    'settings.key' => 'required',
-                    'settings.secret' => 'required',
-                    'settings.region' => 'required',
-                    'settings.configuration_set_name' => 'required'
-                ];
-
-            case self::SENDGRID:
-            case self::POSTMARK:
-                return [
-                    'settings.key' => 'required',
-                ];
-
-            case self::MAILGUN:
-                return [
-                    'settings.key' => 'required',
-                    'settings.domain' => 'required',
-                    'settings.zone' => ['required', 'in:US,EU']
-                ];
-
-            case self::MAILJET:
-                return [
-                    'settings.key' => 'required',
-                    'settings.secret' => 'required',
-                    'settings.zone' => ['required', 'in:Default,US'],
-                ];
-
-            default:
-                return [];
-        }
-    }
-
-    public static function resolveValidationMessages($typeId): array
-    {
-        switch ($typeId) {
-            case self::SES:
-                return [
-                    'settings.key.required' => __('The AWS Email Service requires you to enter a key'),
-                    'settings.secret.required' => __('The AWS Email Service requires you to enter a secret'),
-                    'settings.region.required' => __('The AWS Email Service requires you to enter a region'),
-                    'settings.configuration_set_name.required' => __('The AWS Email Service requires you to enter a configuration set name'),
-                ];
-
-            case self::SENDGRID:
-                return [
-                    'settings.key.required' => __('The Sendgrid Email Service requires you to enter a key'),
-                ];
-
-            case self::POSTMARK:
-                return [
-                    'settings.key.required' => __('The Postmark Email Service requires you to enter a key'),
-                ];
-
-            case self::MAILGUN:
-                return [
-                    'settings.key.required ' =>  __('The Mailgun Email Service requires you to enter a key'),
-                    'settings.domain.required' => __('The Mailgun Email Service requires you to enter a domain'),
-                    'settings.zone.required' => __('The Mailgun Email Service requires you to enter a zone'),
-                ];
-
-            case self::MAILJET:
-                return [
-                    'settings.key.required' =>  __('The Mailjet Email Service requires you to enter a key'),
-                    'settings.secret.required' =>  __('The Mailgun Email Service requires you to enter a secret'),
-                    'settings.zone.required' =>  __('The Mailgun Email Service requires you to enter a zone'),
-                ];
-
-            default:
-                return [];
-        }
-    }
 }
